+++ KestrelMock.Tests/appsettings.json
@@ -17,6 +17,21 @@
 		},
 		{
 			"Request": {
+				"Methods": [ "GET" ],
+				"PathMatchesRegex": ".+\\d{4}.+"
+			},
+			"Response": {
+				"Status": 200,
+				"Headers": [
+					{
+						"Content-Type": "application/json"
+					}
+				],
+				"Body": "{\"banana_x\": 8000}"
+			}
+		},
+		{
+			"Request": {
 				"Methods": [ "POST", "GET" ],
 				"Path": "/hello/world"
 			},
@@ -91,5 +106,6 @@
 				],
 				"Body": "Bad Gateway! Bad!"
 			}
-	}	]
+		}
+	]
 }
